LayoutTests/media/modern-media-controls/invalid-placard/invalid-placard.html

<script src="../../../resources/js-test-pre.js"></script>
<script src="../resources/media-controls-loader.js" type="text/javascript"></script>
<body>
<script type="text/javascript">

description("Testing the <code>InvalidPlacard</code> class.");

const placard = new InvalidPlacard({ layoutTraits: LayoutTraits.macOS });
shouldBe("placard.children[0].children[0].iconName", "Icons.InvalidPlacard");
shouldBe("placard.children[0].children.length", "1");
debug("");

</script>
<script src="../../../resources/js-test-post.js"></script>
</body>
