Safari-611.1.10/LayoutTests/media/modern-media-controls/invalid-placard/invalid-placard-expected.txt

Testing the InvalidPlacard class.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS placard.children[0].children[0].iconName is Icons.InvalidPlacard
PASS placard.children[0].children.length is 1

PASS successfullyParsed is true

TEST COMPLETE
